separate graphs app
