+++ hashtagsv2/graphs/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class GraphsConfig(AppConfig):
+    name = 'graphs'
