+++ hashtagsv2/settings/base.py
@@ -20,6 +20,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'hashtagsv2.hashtags',
+    'hashtagsv2.graphs',
     'django_nose',
 ]
